+++ crates/test/src/lib.rs
@@ -12,12 +12,21 @@ pub use wasm_bindgen_test_macro::wasm_bindgen_test;
 #[global_allocator]
 static A: gg_alloc::GgAlloc<std::alloc::System> = gg_alloc::GgAlloc::new(std::alloc::System);
 
+/// Helper macro which acts like `println!` only routes to `console.error`
+/// instead.
+#[macro_export]
+macro_rules! console_error {
+    ($($arg:tt)*) => (
+        $crate::__rt::console_error(&format_args!($($arg)*))
+    )
+}
+
 /// Helper macro which acts like `println!` only routes to `console.log`
 /// instead.
 #[macro_export]
 macro_rules! console_log {
     ($($arg:tt)*) => (
-        $crate::__rt::log(&format_args!($($arg)*))
+        $crate::__rt::console_log(&format_args!($($arg)*))
     )
 }
 
@@ -87,3 +96,9 @@ macro_rules! wasm_bindgen_test_configure {
 
 #[path = "rt/mod.rs"]
 pub mod __rt;
+
+// Make this only available to wasm32 so that we don't
+// import minicov on other archs.
+// That way you can use normal cargo test without minicov
+#[cfg(target_arch = "wasm32")]
+mod coverage;