+++ src/convert/closures.rs
@@ -54,6 +54,7 @@ macro_rules! stack_closures {
             where $($var: FromWasmAbi,)*
                   R: ReturnWasmAbi
         {
+            #[cfg_attr(wasm_bindgen_unstable_test_coverage, coverage(off))]
             fn describe() {
                 inform(FUNCTION);
                 inform($invoke::<$($var,)* R> as u32);
@@ -108,6 +109,7 @@ macro_rules! stack_closures {
             where $($var: FromWasmAbi,)*
                   R: ReturnWasmAbi
         {
+            #[cfg_attr(wasm_bindgen_unstable_test_coverage, coverage(off))]
             fn describe() {
                 inform(FUNCTION);
                 inform($invoke_mut::<$($var,)* R> as u32);
@@ -151,6 +153,7 @@ where
 }
 
 #[allow(non_snake_case)]
+#[cfg_attr(wasm_bindgen_unstable_test_coverage, coverage(off))]
 unsafe extern "C" fn invoke1_ref<A: RefFromWasmAbi, R: ReturnWasmAbi>(
     a: usize,
     b: usize,
@@ -177,6 +180,7 @@ where
     A: RefFromWasmAbi,
     R: ReturnWasmAbi,
 {
+    #[cfg_attr(wasm_bindgen_unstable_test_coverage, coverage(off))]
     fn describe() {
         inform(FUNCTION);
         inform(invoke1_ref::<A, R> as u32);
@@ -206,6 +210,7 @@ where
 }
 
 #[allow(non_snake_case)]
+#[cfg_attr(wasm_bindgen_unstable_test_coverage, coverage(off))]
 unsafe extern "C" fn invoke1_mut_ref<A: RefFromWasmAbi, R: ReturnWasmAbi>(
     a: usize,
     b: usize,
@@ -232,6 +237,7 @@ where
     A: RefFromWasmAbi,
     R: ReturnWasmAbi,
 {
+    #[cfg_attr(wasm_bindgen_unstable_test_coverage, coverage(off))]
     fn describe() {
         inform(FUNCTION);
         inform(invoke1_mut_ref::<A, R> as u32);